ShowTooltipOnHover implements IPointer*Handler
IPointerClick, IPointerDown and IPointerUp handlers propagate the events to the parent's components implementing the corresponding intefaces, so that ShowTooltipOnHover's event handler doesn't block click events on the parent. (For example, in a dropbox element, if the ShowTooltipOnHover component is placed on the Label, the dropdown doesn't activate if we click on the label itself. This change fixes that unexpected behaviour)