Fixed silly error with cubes' initialScale
Cubes were previously being set to active *after* changing their scale and setting it to the calculated squareSide; this caused the ScaleUpObject's initialScale to be set to that value instead of the intended value, that is the original scale of the prefab.
The scaling mechanism should now be independent from the selected circumference radius and/or the number of cubes (i.e. half of the FFT size).