Created and implemented IReducedBandsSpectrumProvider
Implementers should provide a version of the spectrum data with a reduced number of bands.

My basic implementation divides the spectrum in eight base bands and then subdivides them in the necessary number of subbands.
The maximum number of subbands is limited by a quarter of the FftSize, i.e. half of the number of original bands.