Added a new spectrum visualizer to Visualizer01
The new visualizer uses the recently created IReducedBandsSpectrumProvider to show a spectrum with a reduced number of bands.
The frequency-cubes are placed upon an ellipse's arc, created from a set of customizable parameters such as the maximum length of the arc's chord and the arc extension in degrees.

NOTE: the SpawnEllipseCubes method is quite expensive and, currently, may fail with an exception for certain combinations of parameter values.