Replace the PlayerPrefs-based settings system with a custom one
The new system is based on a hierarchy of ScriptableObject derived classes; each class represents a type of settings (CaptureSettings, QualitySettings, PostProcessingSettings, etc.) with its own set of options.

This system will allow the definition of settings presets which can be swapped and tweaked in real-time in the Editor.
Settings can also be saved to a file on the filesystem and loaded at startup. (Serialization and deserialization is handled with Newtonsoft.Json library).